requirements.txt
Flask~=2.3.3
Pillow~=10.0.0
pyTelegramBotAPI~=4.0.1
hmtai
soundfile~=0.11.0
SpeechRecognition~=3.10.0
certifi~=2024.7.4
requests~=2.32.3
urllib3~=2.2.2
idna~=3.7
